Improve type-casting in `app.rs`.
This commit replaces the repeated `index as usize` with one instance of
shadowing to improve readability. Instances of `as f64` have been
replaced with `.into()` for more explicit casting. Match statements
equating to index values of enums using an `f32` have been amended to
cast the float using `.to_int_unchecked()` to improve readability and
clarity of match cases.