Fix appearance of Slint UI.
This commit sets the Slint UI to compile with the `cupertino-dark` style
in `build.rs`. Having one set style should mitigate concerns with the
application appearing differently on different platforms.

This commit also edits padding and spacing of layouts (still subject to
changes) and changes linear gradient spacing from `20%, 90%` to `0%, 100%`.
This fixes an issue where the Slint VSCode extension live preview would
properly fill the remaining color for the remainder of the gradient,
while the compiled executable would leave undesirable transparent space
in objects and backgrounds.